+++ qa/evil-tests/src/test/java/org/opensearch/common/logging/EvilLoggerTests.java
@@ -62,7 +62,6 @@ import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
-import java.util.Set;
 import java.util.concurrent.BrokenBarrierException;
 import java.util.concurrent.CyclicBarrier;
 import java.util.regex.Matcher;